move AddHandler, UpdateHandler, RemoveHandler and SelectHandler to whinstone
