+++ jkniv/jkniv-whinstone/jkniv-whinstone-cassandra/src/main/java/net/sf/jkniv/whinstone/cassandra/RepositoryCassandra.java
@@ -44,6 +44,7 @@ import net.sf.jkniv.sqlegance.builder.RepositoryConfig;
 import net.sf.jkniv.sqlegance.builder.SqlContextFactory;
 import net.sf.jkniv.whinstone.CommandAdapter;
 import net.sf.jkniv.whinstone.CommandHandler;
+import net.sf.jkniv.whinstone.CommandHandlerFactory;
 import net.sf.jkniv.whinstone.QueryFactory;
 import net.sf.jkniv.whinstone.Queryable;
 import net.sf.jkniv.whinstone.Repository;
@@ -346,7 +347,7 @@ class RepositoryCassandra implements Repository
         if (isTraceEnabled)
             LOG.trace("Executing [{}] as list command", queryable);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new SelectHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofSelect(this.cmdAdapter);
         List<T> list = handler.with(queryable).with(sql).checkSqlType(SqlType.SELECT).with(handlerException)
                 .with(overloadResultRow).run();
         if (isDebugEnabled)
@@ -361,7 +362,7 @@ class RepositoryCassandra implements Repository
         
         T ret = null;
         Sql sql = sqlContext.getQuery(q.getName());
-        CommandHandler handler = new SelectHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofSelect(this.cmdAdapter);
         List<T> list = handler.with(q).with(sql).checkSqlType(SqlType.SELECT).with(handlerException)
                 .with(overloadResultRow).run();
         if (list.size() > 1)
@@ -422,7 +423,7 @@ class RepositoryCassandra implements Repository
     {
         NOT_NULL.verify(queryable);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new AddHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofAdd(this.cmdAdapter);
         int rows = handler.with(queryable).with(sql).checkSqlType(SqlType.INSERT).with(handlerException).run();
         return rows;
     }
@@ -437,8 +438,12 @@ class RepositoryCassandra implements Repository
         
         Queryable queryable = QueryFactory.of(queryName, entity);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new AddHandler(this.cmdAdapter);
-        handler.with(queryable).with(sql).checkSqlType(SqlType.INSERT).with(handlerException).run();
+        CommandHandler handler = CommandHandlerFactory.ofAdd(this.cmdAdapter);
+        handler
+            .with(queryable)
+            .with(sql)
+            .checkSqlType(SqlType.INSERT)
+            .with(handlerException).run();
         return entity;
     }
     
@@ -449,7 +454,7 @@ class RepositoryCassandra implements Repository
         if (isTraceEnabled)
             LOG.trace("Executing [{}] as update command", queryable);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new UpdateHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofUpdate(this.cmdAdapter);
         int rows = handler.with(queryable).with(sql).checkSqlType(SqlType.UPDATE).with(handlerException).run();
         return rows;
     }
@@ -463,7 +468,7 @@ class RepositoryCassandra implements Repository
         if (isTraceEnabled)
             LOG.trace("Executing [{}] as update command", queryable);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new UpdateHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofUpdate(this.cmdAdapter);
         int rows = handler.with(queryable).with(sql).checkSqlType(SqlType.UPDATE).with(handlerException).run();
         return entity;
     }
@@ -580,7 +585,7 @@ class RepositoryCassandra implements Repository
     {
         NOT_NULL.verify(queryable);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new RemoveHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofRemove(this.cmdAdapter);
         int rows = handler.with(queryable).with(sql).checkSqlType(SqlType.DELETE).with(handlerException).run();
         return rows;
     }
@@ -594,7 +599,7 @@ class RepositoryCassandra implements Repository
             LOG.trace("Executing [{}] as remove command", queryName);
         Queryable queryable = QueryFactory.of(queryName, entity);
         Sql sql = sqlContext.getQuery(queryable.getName());
-        CommandHandler handler = new RemoveHandler(this.cmdAdapter);
+        CommandHandler handler = CommandHandlerFactory.ofRemove(this.cmdAdapter);
         int rows = handler.with(queryable).with(sql).checkSqlType(SqlType.DELETE).with(handlerException).run();
         return rows;
     }
