--- jkniv/jkniv-whinstone/jkniv-whinstone-couchdb/src/main/java/net/sf/jkniv/whinstone/couchdb/SelectHandler.java
@@ -1,29 +1,0 @@
-package net.sf.jkniv.whinstone.couchdb;
-
-import net.sf.jkniv.whinstone.Command;
-import net.sf.jkniv.whinstone.CommandAdapter;
-import net.sf.jkniv.whinstone.DefaultQueryHandler;
-
-/**
- * Couchdb Command to handler the {@code Select} life-cycle.
- * 
- * @author Alisson Gomes
- * @since 0.6.0
- */
-class SelectHandler extends DefaultQueryHandler
-{
-    public SelectHandler(CommandAdapter cmdAdapter)
-    {
-        super(cmdAdapter);
-        //with(this);
-    }
-    
-    @Override
-    public Command asCommand()
-    {
-        Command c = getCommandAdapter().asSelectCommand(queryable, overloadResultRow);
-        c.with(this);
-        c.with(this.handleableException);
-        return c;
-    }
-}
